+++ priv/templates/brando.install/lib/application_name_web/endpoint.ex
@@ -24,9 +24,7 @@ defmodule <%= application_module %>Web.Endpoint do
     at: "/",
     from: :<%= application_name %>,
     gzip: true,
-    only: ~w(css fonts img images js ico favicon.ico),
-    cache_control_for_etags: "public, max-age=31536000",
-    cache_control_for_vsn_requests: "public, max-age=31536000"
+    only: ~w(css fonts img images js ico favicon.ico)
 
   plug Plug.Static,
     at: "/media",